# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HelpDesk Pro API - Multitenant support platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tickets:
    get:
      operationId: listTickets
      tags: [tickets]
      summary: List tickets
      security:
        - cookieAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
        - name: tenantId
          in: query
          schema:
            type: integer
        - name: assignedToId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Paginated list of tickets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketListResponse"
    post:
      operationId: createTicket
      tags: [tickets]
      summary: Create a new ticket
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTicketRequest"
      responses:
        "201":
          description: Ticket created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /tickets/{ticketId}:
    get:
      operationId: getTicket
      tags: [tickets]
      summary: Get ticket by ID
      security:
        - cookieAuth: []
      parameters:
        - name: ticketId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ticket details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketDetail"
    patch:
      operationId: updateTicket
      tags: [tickets]
      summary: Update ticket
      security:
        - cookieAuth: []
      parameters:
        - name: ticketId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTicketRequest"
      responses:
        "200":
          description: Ticket updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /tickets/{ticketId}/assign:
    post:
      operationId: assignTicket
      tags: [tickets]
      summary: Assign ticket to technician
      security:
        - cookieAuth: []
      parameters:
        - name: ticketId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignTicketRequest"
      responses:
        "200":
          description: Ticket assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /tickets/{ticketId}/status:
    post:
      operationId: changeTicketStatus
      tags: [tickets]
      summary: Change ticket status
      security:
        - cookieAuth: []
      parameters:
        - name: ticketId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeStatusRequest"
      responses:
        "200":
          description: Status changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /tickets/{ticketId}/comments:
    get:
      operationId: listTicketComments
      tags: [comments]
      summary: List ticket comments
      security:
        - cookieAuth: []
      parameters:
        - name: ticketId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comments list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: addTicketComment
      tags: [comments]
      summary: Add comment to ticket
      security:
        - cookieAuth: []
      parameters:
        - name: ticketId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentRequest"
      responses:
        "201":
          description: Comment added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
...
    Ticket:
      type: object
      properties:
        id:
          type: integer
        ticketNumber:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [nuevo, pendiente, en_revision, en_proceso, esperando_cliente, resuelto, cerrado]
        priority:
          type: string
          enum: [baja, media, alta, urgente]
        category:
          type: string
          nullable: true
        tenantId:
          type: integer
        tenantName:
          type: string
        createdById:
          type: integer
        createdByName:
          type: string
        assignedToId:
          type: integer
          nullable: true
        assignedToName:
          type: string
          nullable: true
        customFields:
          type: object
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
          nullable: true
        commentCount:
          type: integer
      required:
        - id
        - ticketNumber
        - title
        - description
        - status
        - priority
        - tenantId
        - tenantName
        - createdById
        - createdByName
        - createdAt
        - updatedAt
        - commentCount
    TicketDetail:
      allOf:
        - $ref: "#/components/schemas/Ticket"
        - type: object
          properties:
            comments:
              type: array
              items:
                $ref: "#/components/schemas/Comment"
            auditLogs:
              type: array
              items:
                $ref: "#/components/schemas/AuditLog"
    CreateTicketRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 3
        description:
          type: string
          minLength: 10
        priority:
          type: string
          enum: [baja, media, alta, urgente]
          default: media
        category:
          type: string
          nullable: true
        tenantId:
          type: integer
        customFields:
          type: object
          nullable: true
      required:
        - title
        - description
        - tenantId
    UpdateTicketRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [baja, media, alta, urgente]
        category:
          type: string
          nullable: true
        customFields:
          type: object
          nullable: true
    AssignTicketRequest:
      type: object
      properties:
        userId:
          type: integer
          nullable: true
      required:
        - userId
    ChangeStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [nuevo, pendiente, en_revision, en_proceso, esperando_cliente, resuelto, cerrado]
        comment:
          type: string
          nullable: true
      required:
        - status
    TicketListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Ticket"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - data
        - total
        - page
        - limit
        - totalPages
    Comment:
      type: object
      properties:
        id:
          type: integer
        ticketId:
          type: integer
        authorId:
          type: integer
        authorName:
          type: string
        authorRole:
          type: string
        content:
          type: string
        isInternal:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - ticketId
        - authorId
        - authorName
        - authorRole
        - content
        - isInternal
        - createdAt
    CreateCommentRequest:
      type: object
      properties:
        content:
          type: string
          minLength: 1
        isInternal:
          type: boolean
          default: false
      required:
        - content
...
    AuditLog:
      type: object
      properties:
        id:
          type: integer
        action:
          type: string
        entityType:
          type: string
        entityId:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        tenantId:
          type: integer
          nullable: true
        oldValues:
          type: object
          nullable: true
        newValues:
          type: object
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - action
        - entityType
        - entityId
        - userId
        - userName
        - createdAt
...