# lib/api-zod/src/generated/types/changeStatusRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HelpDesk Pro API - Multitenant support platform
 * OpenAPI spec version: 0.1.0
 */
import type { ChangeStatusRequestStatus } from "./changeStatusRequestStatus";

export interface ChangeStatusRequest {
  status: ChangeStatusRequestStatus;
  comment?: string | null;
}