# lib/api-zod/src/generated/types/ticketDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HelpDesk Pro API - Multitenant support platform
 * OpenAPI spec version: 0.1.0
 */
import type { AuditLog } from "./auditLog";
import type { Comment } from "./comment";
import type { Ticket } from "./ticket";

export type TicketDetail = Ticket & {
  comments?: Comment[];
  auditLogs?: AuditLog[];
};